src/backend/requirements.txt
aiohttp==3.8.3
aiomysql==0.1.1
aiosignal==1.3.1
astroid==2.13.3
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==2.1.1
dill==0.3.6
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
isort==5.11.4
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
orjson==3.8.5
platformdirs==2.6.2
PyJWT==2.6.0
pylint==2.15.10
PyMySQL==1.0.2
toml==0.10.2
tomlkit==0.11.6
uvloop==0.17.0
wrapt==1.14.1
yarl==1.8.2
